+++ source/Application/Controller/Admin/ListComponentAjax.php
@@ -8,7 +8,7 @@
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
 use OxidEsales\Eshop\Core\Str;
-use OxidEsales\EshopCommunity\Internal\Transition\ShopEvents\AfterRequestProcessedEvent;
+use OxidEsales\EshopCommunity\Internal\Transition\ShopEvents\AfterAdminAjaxRequestProcessedEvent;
 
 /**
  * AJAX call processor class
@@ -143,7 +143,7 @@ class ListComponentAjax extends \OxidEsales\Eshop\Core\Base
     {
         if ($function) {
             $this->$function();
-            $this->dispatchEvent(new AfterRequestProcessedEvent());
+            $this->dispatchEvent(new AfterAdminAjaxRequestProcessedEvent);
         } else {
             $sQAdd = $this->_getQuery();