+++ tests/Unit/Core/BaseTest.php
@@ -709,60 +709,43 @@ class BaseTest extends \OxidTestCase
         $this->assertEquals("test class", $oBase->getClassName());
     }
 
-    /**
-     * Test get update fields.
-     *
-     * @return null
-     */
     public function testGetUpdateFields()
     {
-        $oBase = new _oxBase();
-        $oBase->init("oxactions");
-        $oBase->oxactions__oxid = new oxField("test1", oxField::T_RAW);
-        $oBase->oxactions__oxtitle = new oxField("title1", oxField::T_RAW);
+        $base = new _oxBase();
+        $base->init("oxactions");
+        $base->oxactions__oxid = new oxField("test1", oxField::T_RAW);
+        $base->oxactions__oxtitle = new oxField("title1", oxField::T_RAW);
 
-        $shopId = $this->getUpdateShopId();
-        $sGetUpdateFields = "oxid = 'test1',oxshopid = '{$shopId}',oxtype = '',oxtitle = 'title1',oxtitle_1 = '',oxtitle_2 = '',oxtitle_3 = '',oxlongdesc = '',oxlongdesc_1 = '',oxlongdesc_2 = '',oxlongdesc_3 = '',oxactive = '1',oxactivefrom = '0000-00-00 00:00:00',oxactiveto = '0000-00-00 00:00:00',oxpic = '',oxpic_1 = '',oxpic_2 = '',oxpic_3 = '',oxlink = '',oxlink_1 = '',oxlink_2 = '',oxlink_3 = '',oxsort = '0'";
-        $this->assertEquals($sGetUpdateFields, $oBase->UNITgetUpdateFields());
+        $expectedOxid = "oxid = 'test1'";
+        $this->assertStringStartsWith($expectedOxid, $base->UNITgetUpdateFields());
     }
 
-    /**
-     * Test get update fields with use skip save fields of.
-     *
-     * @return null
-     */
     public function testGetUpdateFieldsWithUseSkipSaveFieldsOff()
     {
-        $oBase = new _oxBase();
-        $oBase->init("oxactions");
-        $oBase->oxactions__oxid = new oxField("test1", oxField::T_RAW);
-        $oBase->oxactions__oxtitle = new oxField("title1", oxField::T_RAW);
-
-        $oBase->setClassVar('_aSkipSaveFields', array('oxtitle'));
+        $base = new _oxBase();
+        $base->init('oxactions');
+        $base->oxactions__oxid = new oxField('test1', oxField::T_RAW);
+        $base->oxactions__oxtitle = new oxField('title1', oxField::T_RAW);
+        $base->setClassVar('_aSkipSaveFields', ['oxtitle']);
 
         $shopId = $this->getUpdateShopId();
-        $sGetUpdateFields = "oxid = 'test1',oxshopid = '{$shopId}',oxtype = '',oxtitle = 'title1',oxtitle_1 = '',oxtitle_2 = '',oxtitle_3 = '',oxlongdesc = '',oxlongdesc_1 = '',oxlongdesc_2 = '',oxlongdesc_3 = '',oxactive = '1',oxactivefrom = '0000-00-00 00:00:00',oxactiveto = '0000-00-00 00:00:00',oxpic = '',oxpic_1 = '',oxpic_2 = '',oxpic_3 = '',oxlink = '',oxlink_1 = '',oxlink_2 = '',oxlink_3 = '',oxsort = '0',oxtimestamp = 'CURRENT_TIMESTAMP'";
-        $this->assertEquals($sGetUpdateFields, $oBase->UNITgetUpdateFields(false));
+
+        $expectedOxid = "oxid = 'test1',oxshopid = '{$shopId}',oxtype = '',oxtitle = 'title1'";
+        $this->assertStringStartsWith($expectedOxid, $base->UNITgetUpdateFields(false));
     }
 
-    /**
-     * Test get update fields with use skip save fields on.
-     *
-     * @return null
-     */
     public function testGetUpdateFieldsWithUseSkipSaveFieldsOn()
     {
-        $oBase = new _oxBase();
-        $oBase->init("oxactions");
-        $oBase->oxactions__oxid = new oxField("test1", oxField::T_RAW);
-        $oBase->oxactions__oxtitle = new oxField("title1", oxField::T_RAW);
-
-        $oBase->setClassVar('_aSkipSaveFields', array('oxtitle'));
+        $base = new _oxBase();
+        $base->init('oxactions');
+        $base->oxactions__oxid = new oxField('test1', oxField::T_RAW);
+        $base->oxactions__oxtitle = new oxField('title1', oxField::T_RAW);
 
         $shopId = $this->getUpdateShopId();
-        $sGetUpdateFields = "oxid = 'test1',oxshopid = '{$shopId}',oxtype = '',oxtitle_1 = '',oxtitle_2 = '',oxtitle_3 = '',oxlongdesc = '',oxlongdesc_1 = '',oxlongdesc_2 = '',oxlongdesc_3 = '',oxactive = '1',oxactivefrom = '0000-00-00 00:00:00',oxactiveto = '0000-00-00 00:00:00',oxpic = '',oxpic_1 = '',oxpic_2 = '',oxpic_3 = '',oxlink = '',oxlink_1 = '',oxlink_2 = '',oxlink_3 = '',oxsort = '0',oxtimestamp = 'CURRENT_TIMESTAMP'";
 
-        $this->assertEquals($sGetUpdateFields, $oBase->UNITgetUpdateFields());
+        $base->setClassVar('_aSkipSaveFields', ['oxtitle']);
+        $expectedOxid = "oxid = 'test1',oxshopid = '{$shopId}',oxtype = '',oxtitle_1 = ''";
+        $this->assertStringStartsWith($expectedOxid, $base->UNITgetUpdateFields());
     }
 
     /**